Förbättringar till frontend för att underlätta utveckling

+++ frack/app/routes.py
@@ -7,7 +7,7 @@ import os
 from app.models.news import News
 import app.news_functions as news_functions
 from sqlalchemy import desc
-from app.authentication import requires_auth, requires_auth_token
+from app.authentication import requires_auth, requires_auth_token, check_auth
 
 
 
@@ -15,8 +15,9 @@ STATIC_DIR = os.path.join(os.getcwd(), "static", "Schmeck")
 
 
 #Definiera olika URL-er och vad de leder till
-@app.route("/")
-def index():
+@app.route('/', defaults={'path': ''})
+@app.route('/<path:path>')
+def index(path):
     return send_from_directory(STATIC_DIR, "index.html")
 
 #ladda CSS
@@ -29,52 +30,45 @@ def get_css(filename):
 def get_js(filename):
     return send_from_directory(os.path.join(STATIC_DIR, "js"), filename)
 
+#ladda templates
+@app.route("/templates/<filename>")
+def get_templates(filename):
+    return send_from_directory(os.path.join(STATIC_DIR, "templates"), filename)
+
+@app.route("/images/<filename>")
+def get_images(filename):
+    return send_from_directory(os.path.join(STATIC_DIR, "images"), filename)
+
 #ladda media (bild, film, osv)
 @app.route("/api/media/<file_path>")
 def get_media(file_path):
     return send_from_directory(os.path.join(STATIC_DIR, "media"), file_path)
 
-@app.route("/news")
-def news_page():
-    return send_from_directory(STATIC_DIR, "news.html")
-
-@app.route("/news/<id>")
-def news_page_specific(id):
-    return send_from_directory(STATIC_DIR, "news.html")
-
-@app.route("/news/edit/<id>")
-def edit_page(id):
-    return send_from_directory(STATIC_DIR, "edit.html")
-
-@app.route("/api/news/all")
-def get_news():
-    return news_functions.get_all_news()
-
-@app.route("/api/news/<id>")
-def get_news_by_id(id):
-    return get_news_by_id(id), 201
-
-@app.route("/api/news/upload", methods=["POST"])
+@app.route("/api/news/", methods=["GET", "POST"])
+@app.route("/api/news/<id>", methods=["GET", "DELETE", "PUT"])
 @requires_auth_token
-def add_news():
-    return news_functions.add_news(request.json), 200
-
-
-@app.route("/api/news/delete/<id>")
-@requires_auth_token
-def delete_news(id):
-    return news_functions.delete_news(id), 200
-
-@app.route("/api/news/edit/<id>", methods=["POST"])
-@requires_auth_token
-def edit_news(id):
-    return news_functions.edit_news(id, request.json), 201
+def news_route(id=None):
+    if request.method == "GET":
+        if id == None:
+            return news_functions.get_all_news()
+        else:
+            return news_functions.get_news_by_id(id), 201
+    elif request.method == "POST":
+        return news_functions.add_news(request.json), 200
+    elif request.method == "DELETE":
+        return news_functions.delete_news(id), 200
+    elif request.method == "PUT":
+        return news_functions.edit_news(id, request.json), 201
 
 @app.route('/api/token')
 @requires_auth
 def get_auth_token():
     token = g.user.generate_auth_token()
     return jsonify({ 'token': token.decode('ascii') })
+
+@app.route('/api/checkToken', methods=["POST"])
+def check_token():
+    return jsonify({"valid": check_auth(request.json["token"], "", True)})
 
 @app.route("/api/hemlig")
 @requires_auth_token
